Clarify and fix the semantics of the dates in my `hhs_hospitals_bitemporal` model, that's subtly different from the HHS's hospital utilization file.
The HHS files' `date` column is based on data publication date, as can be seen from the fact that the file published e.g. on April 25 has a field with date of April 25, and columns with names like `previous_day_admission_adult_covid_confirmed` that say their data is about the "previous day."

The semantics of this project—embodied e.g. in the `bulletin_date` column—is that instead of publication date we prefer "data up to" date. A file published on April 25 is presumed to have data as of the closing of April 24.